fix: move logo to assets
