+++ react-ts-app/src/components/Footer/Footer.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import styles from './footer.module.scss';
-import { ReactComponent as Logo } from '../../../public/logo.svg';
+import { ReactComponent as Logo } from '../../assets/images/logo.svg';
 
 class Footer extends Component {
   render() {
