+++ react-ts-app/src/components/Header/Header.tsx
@@ -1,18 +1,19 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 import Navigation from '../Navigation/Navigation';
 import styles from './header.module.scss';
 
-import { ReactComponent as Logo } from '../../../public/logo.svg';
+import { ReactComponent as Logo } from '../../assets/images/logo.svg';
 
 class Header extends React.Component {
   render() {
     return (
       <header className={styles.header}>
         <div className={styles.container}>
-          <div className={styles.logo}>
+          <Link to="/" className={styles.logo}>
             <Logo />
             <span>cultured kid.</span>
-          </div>
+          </Link>
           <Navigation />
         </div>
       </header>
